backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
authlib
python-jose
passlib[bcrypt]
httpx
razorpay
apscheduler
python-dotenv
pydantic